+++ apps/common/src/main/scala/com/daml/network/environment/TopologyAdminConnection.scala
@@ -1462,12 +1462,15 @@ abstract class TopologyAdminConnection(
         member.filterString,
       )
     ).map(
-      _.map(r =>
-        TopologyResult(
-          r.context,
-          r.item,
+      // TODO(#14815) Canton currently compares member IDs by string prefix instead of strict equality of
+      // member IDs in ListDomainTrustCertificate, so we apply another filter for equality of the member ID
+      _.filter(r => r.item.participantId.member.filterString == member.filterString)
+        .map(r =>
+          TopologyResult(
+            r.context,
+            r.item,
+          )
         )
-      )
     )
 
   def ensureDomainTrustCertificateRemoved(