+++ cluster/deployment/compose/start.sh
@@ -21,15 +21,17 @@ function _info(){
 }
 
 function usage() {
-  echo "Usage: $0 -s <sponsor_sv_address> -o <onboarding_secret> -p <party_hint> [-a] [-b] [-c <scan_address>] [-q <sequencer_address>] [-n <network_name>] [-m <migration_id>] [-M]"
+  echo "Usage: $0 -s <sponsor_sv_address> -o <onboarding_secret> -p <party_hint> [-a] [-b] [-c <scan_address>] [-q <sequencer_address>] [-n <network_name>] [-m <migration_id>] [-M] [-i <identities_dump>] [-P <participant_id>]"
   echo "  -s <sponsor_sv_address>: The full URL of the sponsor SV"
   echo "  -o <onboarding_secret>: The onboarding secret to use. If not provided, it will be fetched from the sponsor SV (possible on DevNet only)"
-  echo "  -p <party_hint>: The party hint to use for the validator operator, will also act as the participant identifier."
+  echo "  -p <party_hint>: The party hint to use for the validator operator, by default also your participant identifier."
+  echo "  -P <participant_id>: The participant identifier."
   echo "  -a: Use this flag to enable authentication"
   echo "  -c <scan_address>: The full URL of a Scan app. If not provided, it will be derived from the sponsor SV address."
   echo "  -n <network_name>: The name of an existing docker network to use. If not provided, the default network will be created used."
   echo "  -m <migration_id>: The migration ID to use. Must be a non-negative integer."
   echo "  -M: Use this flag when bumping the migration ID as part of a migration."
+  echo "  -i <identities_dump>: restore identities from a dump file. Requires a new participant identifier to be provided."
 
   echo ""
   echo "Testing flags:"
@@ -51,7 +53,9 @@ network_name=""
 migration_id=0
 migrating=0
 party_hint=""
-while getopts 'has:c:t:o:n:bq:m:Mp:' arg; do
+participant_id=""
+restore_identities_dump=""
+while getopts 'has:c:t:o:n:bq:m:Mp:P:i:' arg; do
   case ${arg} in
     h)
       usage
@@ -86,6 +90,12 @@ while getopts 'has:c:t:o:n:bq:m:Mp:' arg; do
       ;;
     p)
       party_hint="${OPTARG}"
+      ;;
+    P)
+      participant_id="${OPTARG}"
+      ;;
+    i)
+      restore_identities_dump="${OPTARG}"
       ;;
     ?)
       usage
@@ -129,13 +139,23 @@ if [[ ! "${migration_id}" =~ ^[0-9]+$ ]]; then
   exit 1
 fi
 
+if [ -n "${restore_identities_dump}" ] && [ -z "${participant_id}" ]; then
+  _error_msg "Please provide a new participant identifier when restoring identities from a dump file"
+  usage
+  exit 1
+fi
+
 export ONBOARDING_SECRET
 export SPONSOR_SV_ADDRESS
 export SCAN_ADDRESS
 export SEQUENCER_ADDRESS
 export MIGRATION_ID=${migration_id}
-export PARTICIPANT_IDENTIFIER=${party_hint}
 export PARTY_HINT=${party_hint}
+if [ -n "${participant_id}" ]; then
+  export PARTICIPANT_IDENTIFIER=${participant_id}
+else
+  export PARTICIPANT_IDENTIFIER=${party_hint}
+fi
 
 # TODO(#14303): release tag should be injected by the release pipeline
 # IMAGE_TAG=$("${REPO_ROOT}/build-tools/get-snapshot-version")
@@ -176,5 +196,10 @@ if [ -n "${network_name}" ]; then
   # it declares the same name as the network_name argument.
   extra_compose_files+=("-f" "${script_dir}/compose-onvpn-network.yaml")
   export DOCKER_NETWORK="${network_name}"
+fi
+if [ -n "${restore_identities_dump}" ]; then
+  extra_compose_files+=("-f" "${script_dir}/compose-restore-from-id.yaml")
+  export VALIDATOR_NEW_PARTICIPANT_IDENTIFIER=${PARTICIPANT_IDENTIFIER}
+  export VALIDATOR_PARTICIPANT_IDENTITIES_DUMP=${restore_identities_dump}
 fi
 docker compose -f "${script_dir}/compose.yaml" "${extra_compose_files[@]}" up -d